changed stat to float from int in items

+++ Assets/Scripts/Classes/Item.cs
@@ -6,7 +6,7 @@ public class Item
 	public string Name { get; private set; }
 	public int ID { get; private set; }
 	public string Description { get; private set; }
-	public Dictionary<string, int> Stats { get; private set; } // enhancements it provides methinks
+	public Dictionary<string, float> Stats { get; private set; } // enhancements it provides methinks
 	public int Value { get; private set; } // how much its worth when buying
 	public int SellValue { get; private set; } // how much its worth when selling
 
@@ -15,12 +15,12 @@ public class Item
 		Name = "[UNINITIALIZED ITEM]";
 		ID = -1;
 		Description = "[UNINITIALIZED DESCRIPTION]";
-		Stats = new Dictionary<string, int>();
+		Stats = new Dictionary<string, float>();
 		Value = 0;
 		SellValue = 0;
 	}
 
-	public Item(string name, int id, string description, Dictionary<string, int> stats, int value, int sellValue)
+	public Item(string name, int id, string description, Dictionary<string, float> stats, int value, int sellValue)
 	{
 		this.Name = name;
 		this.ID = id;
@@ -33,7 +33,7 @@ public class Item
 	public override string ToString()
 	{
 		string returnStr = "ITEM PRINT\nName: " + Name + "\nID: " + ID + "\nDescription: " + Description + "\nValue: " + Value + "\nSell Value: " + SellValue + "\nStats: ";
-		foreach (KeyValuePair<string, int> pair in Stats)
+		foreach (KeyValuePair<string, float> pair in Stats)
 		{
 			returnStr += "(";
 			returnStr += pair.Key;
